complete 4.21

+++ osa4/controllers/blogs.js
@@ -16,7 +16,7 @@ const jwt = require('jsonwebtoken')
   
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     
-
+  
     if (!request.token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
     }
@@ -45,9 +45,24 @@ const jwt = require('jsonwebtoken')
   })
 
   blogsRouter.delete('/blogs/:id', async (request, response) => {
+    
+    const decodedToken =  jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
 
+    if(!decodedToken){
+      return response.status(401).json({ error: 'token missing or invalid' })
+      }
+   
+    if(blog.user.toString() === user.id.toString()){
     await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    return response.status(204).end()
+    }
+    if(blog.user.toString() !== user.id.toString()){
+      return response.status(401).json({error: 'Not authorized'})
+    }
+
+    
   })
 
   blogsRouter.put('/blogs/:id', async (request, response) => {
